backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
eventlet==0.33.3
psycopg2-binary==2.9.7
gunicorn==21.2.0